requirements.txt
click==7.1.2
fid-score==0.1.3
h5py==2.10.0
imageio==2.6.1
imgaug==0.4.0
Markdown==3.1.1
matplotlib==3.3.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.5.5.64
pandas==1.1.5
Pillow==6.2.1
protobuf==3.19.6
pytorch-fid==0.2.0
pytorch-lightning==1.2.9
pytorch3d==0.3.0
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==23.1.0
scikit-image==0.17.2
scipy==1.3.2
six==1.16.0
tensorboardX==2.5
timm==0.4.12
toml==0.10.1
toolz==0.10.0
torch==1.4.0
torchmetrics==0.2.0
torchtext==0.6.0
torchvision==0.5.0
tornado==6.1
tqdm==4.62.3
zipp==3.3.1
